Thank you. I found and fixed two show-stopper bugs:
1. Google changed the format of returned results to omit the word
"pixels" in the image sizes, which basically made it so that
RandomWeb wouldn't actually download anything.

2. RandomWeb was loading a "search size" preference, which I assume
was a relic of old releases, since there is no UI to control this in
the preferences. Anyway, it was causing big problems with image
searching (no results were being returned), so I disabled loading
that preference.

Patch from Nick Zitzmann